requirements.txt
openai>=1.3.0
google-cloud-aiplatform>=1.38.0
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
nltk>=3.8.0
pandas>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 